+++ openlevelup/js/model.js
@@ -177,7 +177,8 @@ MapData: function(ctrl) {
 			//Edit indoor areas to set them as polygons instead of linestrings
 			if(((feature.properties.tags.indoor != undefined
 				&& feature.properties.tags.indoor != "yes")
-				|| feature.properties.tags.buildingpart != undefined)
+				|| feature.properties.tags.buildingpart != undefined
+				|| feature.properties.tags.highway == "elevator")
 				&& feature.geometry.type == "LineString") {
 				
 				feature = _convertLineToPolygon(feature);